+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `.shulkr` file for repository settings (currently only saves mappings)
+- `.shulkr` file for repository settings (currently only saves mappings and decompiler)
 
 ### Changed
 - Mojang's official mappings are now used by default